+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/entities/TileRepository.kt
@@ -1,6 +1,6 @@
 package de.gleex.pltcmd.game.ui.entities
 
-import de.gleex.pltcmd.model.radio.SignalStrength
+import de.gleex.pltcmd.model.radio.broadcasting.SignalStrength
 import de.gleex.pltcmd.model.world.terrain.Terrain
 import de.gleex.pltcmd.model.world.terrain.TerrainHeight
 import de.gleex.pltcmd.model.world.terrain.TerrainType
