+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/renderers/RadioSignalVisualizer.kt
@@ -4,8 +4,8 @@ import de.gleex.pltcmd.game.options.GameOptions
 import de.gleex.pltcmd.game.ui.entities.GameBlock
 import de.gleex.pltcmd.game.ui.entities.GameWorld
 import de.gleex.pltcmd.game.ui.entities.TileRepository
-import de.gleex.pltcmd.model.radio.AttenuationModel
-import de.gleex.pltcmd.model.radio.RadioSignal
+import de.gleex.pltcmd.model.radio.broadcasting.AttenuationModel
+import de.gleex.pltcmd.model.radio.broadcasting.RadioSignal
 import org.hexworks.cobalt.databinding.api.property.Property
 import org.hexworks.zircon.api.data.Position
 import org.hexworks.zircon.api.data.Size
