+++ model/radio/src/main/kotlin/de/gleex/pltcmd/model/radio/broadcasting/RadioSignal.kt
@@ -1,4 +1,4 @@
-package de.gleex.pltcmd.model.radio
+package de.gleex.pltcmd.model.radio.broadcasting
 
 import de.gleex.pltcmd.model.world.terrain.Terrain
 import de.gleex.pltcmd.model.world.terrain.TerrainHeight
